requirements.txt
chromadb
fastapi
google-genai
langchain-core
langgraph
matplotlib
numpy
pandas
Pillow
pydantic
pytz
qstrader
scipy
yfinance
